# backend/src/wto_backend/contract_data/openapi-bundled.json
{
  "components": {
    "headers": {
      "CorrelationId": {
        "description": "Request correlation identifier",
        "required": true,
        "schema": {
          "maxLength": 128,
          "minLength": 1,
          "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$",
          "type": "string"
        }
      }
    },
    "parameters": {
      "AgentId": {
        "in": "path",
        "name": "agent_id",
        "required": true,
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      },
...
      "CredentialId": {
        "in": "path",
        "name": "credential_id",
        "required": true,
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      },
...
    },
...
    "responses": {
      "CredentialMetadataList": {
        "content": {
          "application/json": {
            "schema": {
              "items": {
                "$ref": "#/components/schemas/CredentialMetadata"
              },
              "maxItems": 128,
              "type": "array"
            }
          }
        },
        "description": "Credential versions without secrets",
        "headers": {
          "X-Correlation-ID": {
            "$ref": "#/components/headers/CorrelationId"
          }
        }
      },
...
      "Error": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        },
        "description": "Normative error",
        "headers": {
          "X-Correlation-ID": {
            "$ref": "#/components/headers/CorrelationId"
          }
        }
      },
...
    },
    "schemas": {
...
      "CredentialMetadata": {
        "additionalProperties": false,
        "properties": {
          "activated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "credential_id": {
            "format": "uuid",
            "type": "string"
          },
          "credential_version": {
            "minimum": 1,
            "type": "integer"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "issued_at": {
            "format": "date-time",
            "type": "string"
          },
          "revocation_reason": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "schema_version": {
            "const": "1.0.0",
            "type": "string"
          },
          "state": {
            "enum": [
              "pending",
              "active",
              "revoked",
              "expired"
            ],
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "credential_id",
          "credential_version",
          "state",
          "issued_at",
          "activated_at",
          "expires_at",
          "revoked_at",
          "revocation_reason"
        ],
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "$id": "https://schemas.wifi-test-orchestrator.example/1.0.0/error-envelope.schema.json",
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "additionalProperties": false,
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "maxLength": 64,
                "minLength": 3,
                "pattern": "^[a-z][a-z0-9_]+$",
                "type": "string"
              },
              "correlation_id": {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$",
                "type": "string"
              },
              "details": {
                "anyOf": [
                  {
                    "items": {
                      "additionalProperties": false,
                      "properties": {
                        "location": {
                          "items": {
                            "maxLength": 64,
                            "type": "string"
                          },
                          "maxItems": 8,
                          "type": "array"
                        },
                        "type": {
                          "maxLength": 64,
                          "type": "string"
                        }
                      },
                      "required": [
                        "location",
                        "type"
                      ],
                      "type": "object"
                    },
                    "maxItems": 32,
                    "type": "array"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "message": {
                "maxLength": 256,
                "minLength": 1,
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "details",
              "correlation_id"
            ],
            "type": "object"
          },
          "schema_version": {
            "$ref": "#/components/schemas/WTOCommonTypes/$defs/schema_version"
          }
        },
        "required": [
          "schema_version",
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AgentBearerAuth": {
        "bearerFormat": "WTO-AgentCredential-v1",
        "scheme": "bearer",
        "type": "http"
      },
      "AgentPendingCredentialAuth": {
        "bearerFormat": "WTO-PendingAgentCredential-v1",
        "scheme": "bearer",
        "type": "http"
      },
      "UserBearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "UserRefreshCookie": {
        "in": "cookie",
        "name": "wto_refresh",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Canonical Phase 04 API. Internal Phase 03 operations remain explicitly provisional.",
    "title": "WiFi Test Orchestrator API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/agents/{agent_id}/credentials": {
      "get": {
        "operationId": "listAgentCredentials",
        "parameters": [
          {
            "$ref": "#/components/parameters/AgentId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/CredentialMetadataList"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "UserBearerAuth": []
          }
        ],
        "tags": [
          "agent-administration"
        ]
      }
    },
    "/api/v1/agents/{agent_id}/credentials/{credential_id}/revoke": {
      "post": {
        "operationId": "revokeAgentCredential",
        "parameters": [
          {
            "$ref": "#/components/parameters/AgentId"
          },
          {
            "$ref": "#/components/parameters/CredentialId"
          }
        ],
        "responses": {
          "204": {
            "description": "Credential revoked",
            "headers": {
              "X-Correlation-ID": {
                "$ref": "#/components/headers/CorrelationId"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "UserBearerAuth": []
          }
        ],
        "tags": [
          "agent-administration"
        ]
      }
    },
...
    "/api/v1/agents/{agent_id}/revoke": {
      "post": {
        "operationId": "revokeAgent",
        "parameters": [
          {
            "$ref": "#/components/parameters/AgentId"
          }
        ],
        "responses": {
          "204": {
            "description": "Agent revoked",
            "headers": {
              "X-Correlation-ID": {
                "$ref": "#/components/headers/CorrelationId"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "UserBearerAuth": []
          }
        ],
        "tags": [
          "agent-administration"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}